Updated user status

+++ app/User.php
@@ -52,7 +52,7 @@ class User extends Authenticatable
      */
     protected $fillable = [
         'timeline_id', 'email', 'verification_code', 'email_verified', 'remember_token', 'password', 'birthday', 'city', 'gender', 'last_logged', 'timezone', 'affiliate_id', 'language', 'country', 'active', 'verified', 'facebook_link', 'twitter_link', 'dribbble_link', 'instagram_link', 'youtube_link', 'linkedin_link', 'wishlist', 'website', 'instagram','custom_option1', 'custom_option2', 'custom_option3', 'custom_option4'
-        , 'bank_account', 'price', 'is_payment_set', 'is_bank_set', 'is_follow_for_free'
+        , 'bank_account', 'price', 'is_payment_set', 'is_bank_set', 'is_follow_for_free', 'is_online'
     ];
 
     /**
@@ -505,6 +505,7 @@ class User extends Authenticatable
 
     public function deleteOthers()
     {
+        $authUserId = Auth::user()->id;
         $otherposts = $this->timeline->posts()->where('user_id', '!=', Auth::user()->id)->get();
         foreach ($otherposts as $otpost) {
             $otpost->users_liked()->detach();
@@ -525,7 +526,7 @@ class User extends Authenticatable
             // if($otpost->shared_post_id != NULL) {
             //     $otpost->update(['shared_post_id' => null])->save();
             // }
-            if (count($otpost->sharedPost()->first()) != 0) {
+            if (!empty($otpost->sharedPost()->first()) && count($otpost->sharedPost()->first()) != 0) {
                 $otpost->sharedPost->delete();
             }
             $otpost->delete();
